Added integration tests by combing the voltage and current stage PID. Also added colors to the printf statements. Fixed a bug within the saturation function caused by a comparison between float and uint32_t.

+++ src/pid.h
@@ -6,8 +6,8 @@ typedef struct
     float kI;
     float KP;
     float kD;
-    uint32_t upperLimit;
-    uint32_t lowerLimit;
+    float upperLimit;
+    float lowerLimit;
     float error;
     float referencePoint;
     float previousError;
